Fix valueCat string extender reference counts.

+++ js_malloc.h
@@ -5,8 +5,8 @@
 //
 
 typedef struct RawObj {
-	int64_t weakCnt[1];
-	int64_t refCnt[1];
+	int32_t weakCnt[1];
+	int32_t refCnt[1];
 	uint64_t addr;
 } rawobj_t;
